first pass, for the sake of NrrdIO, at resolving signed vs. unsigned comparison warnings
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2468 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/air/sane.c
@@ -91,7 +91,7 @@ airSanity(void) {
   ninfF = ninf;
   airFPValToParts_f(&sign, &exp, &mant, nanF);
   mant >>= 22;
-  if (AIR_QNANHIBIT != mant) {
+  if (AIR_QNANHIBIT != (int)mant) {
     return airInsane_QNaNHiBit;
   }
   if (!(airFP_QNAN == airFPClass_f(nanF)
